# openapi.yaml
openapi: 3.0.3
info:
  title: Content Broadcasting System API
  description: >
    REST API for managing and broadcasting digital content across display
    devices in an educational institution. Teachers upload content; a principal
    reviews, approves, and schedules it into display slots.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token (15 min default) returned on login/refresh.

    csrfHeader:
      type: apiKey
      in: header
      name: x-xsrf
      description: >
        Double-submit CSRF token. Must match the value of the `XSRF-TOKEN`
        cookie set at login/register. Required on all state-changing requests
        (every POST, PATCH, and DELETE that requires authentication).
        Public endpoints and GET requests are exempt.
  schemas:
...
    ContentDto:
      type: object
      properties:
        id:
          type: integer
          example: 7
        title:
          type: string
          example: Morning Maths Slide
        description:
          type: string
          nullable: true
          example: Slide for 8 AM maths period
        subject:
          type: string
          example: mathematics
        file_path:
          type: string
          description: Internal S3 object key. Use file_url for direct access.
          example: content/2025/01/uuid.jpg
        file_type:
          type: string
          example: image/jpeg
        file_size:
          type: integer
          description: File size in bytes.
          example: 204800
        uploaded_by:
          type: integer
          example: 3
        status:
          type: string
          enum: [pending, approved, rejected]
          example: pending
        rejection_reason:
          type: string
          nullable: true
          example: Image resolution too low
        approved_by:
          type: integer
          nullable: true
          example: 1
        approved_at:
          type: string
          format: date-time
          nullable: true
        start_time:
          type: string
          format: date-time
          nullable: true
          example: "2025-06-01T08:00:00Z"
        end_time:
          type: string
          format: date-time
          nullable: true
          example: "2025-06-30T17:00:00Z"
        rotation_duration_minutes:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1440
          example: 10
        created_at:
          type: string
          format: date-time
        file_url:
          type: string
          description: Pre-signed S3 GET URL, valid for 1 hour.
          example: "https://bucket.s3.region.amazonaws.com/content/2025/01/uuid.jpg?X-Amz-Signature=..."
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        pages:
          type: integer
          example: 3
    PaginatedContent:
      type: object
      properties:
        success:
          type: boolean
          example: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContentDto'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Human-readable error description
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication required
    Forbidden:
      description: Authenticated but insufficient role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Forbidden: insufficient permissions"
    NotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Content not found
    BadRequest:
      description: Validation error or malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "title: Required; subject: Required"
...
paths:
...
  /content:
    post:
      tags: [Content]
      summary: Upload content (teacher)
      description: >
        Uploads an image file to S3 and creates a content record with status
        `pending`. Only teachers can upload. The file is validated by MIME type
        (jpg/png/gif) and size (max 10 MB).
      security:
        - bearerAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title, subject]
              properties:
                file:
                  type: string
                  format: binary
                  description: 'Image file. Allowed types: image/jpeg, image/png, image/gif. Max 10 MB.'
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  example: Morning Maths Slide
                subject:
                  type: string
                  minLength: 1
                  maxLength: 80
                  example: mathematics
                description:
                  type: string
                  maxLength: 2000
                  nullable: true
                  example: Used during the 8 AM period
                start_time:
                  type: string
                  format: date-time
                  description: ISO 8601. Content is only active on or after this time.
                  example: "2025-06-01T08:00:00Z"
                end_time:
                  type: string
                  format: date-time
                  description: ISO 8601. Must be after start_time. Content stops being active after this time.
                  example: "2025-06-30T17:00:00Z"
                rotation_duration_minutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                  description: Suggested display duration per rotation cycle.
                  example: 10
      responses:
        '201':
          description: Content uploaded and record created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  content:
                    $ref: '#/components/schemas/ContentDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags: [Content]
      summary: List all content (principal)
      description: Returns a paginated list of all content across all teachers, with optional filters.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
          description: Filter by approval status.
        - in: query
          name: subject
          schema:
            type: string
            maxLength: 80
          description: Filter by subject (case-insensitive).
        - in: query
          name: uploaded_by
          schema:
            type: integer
          description: Filter by teacher user ID.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated content list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /content/mine:
    get:
      tags: [Content]
      summary: List own uploads (teacher)
      description: Returns a paginated list of content uploaded by the authenticated teacher.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
          description: Filter by approval status.
        - in: query
          name: subject
          schema:
            type: string
            maxLength: 80
          description: Filter by subject (case-insensitive).
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated content list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /content/{id}/approve:
    patch:
      tags: [Content]
      summary: Approve content (principal)
      description: >
        Sets content status to `approved`. Only principals can call this.
        Approved content can be scheduled for display.
      security:
        - bearerAuth: []
          csrfHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Content ID.
      responses:
        '200':
          description: Content approved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  content:
                    $ref: '#/components/schemas/ContentDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/{id}/reject:
    patch:
      tags: [Content]
      summary: Reject content (principal)
      description: >
        Sets content status to `rejected` with a mandatory reason.
        Rejected content cannot be scheduled.
      security:
        - bearerAuth: []
          csrfHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Content ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rejection_reason]
              properties:
                rejection_reason:
                  type: string
                  minLength: 1
                  maxLength: 500
                  example: Image resolution too low for display screens
      responses:
        '200':
          description: Content rejected.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  content:
                    $ref: '#/components/schemas/ContentDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...